pyproject.toml
[tool.poetry]
name = "nowcasting_gdp"
version = "0.1.0"
description = "Nowcasting pipeline for macroeconomic analysis"
authors = ["Renato Vassallo <renato.vassallo@bse.eu>"]
package-mode = false  
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pmdarima = "^2.0.4"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.2"
fredapi = "^0.5.2"
numpy = "^2.1.2"
xgboost = "^2.1.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
